+++ src/main/java/com/kakaotech/team18/backend_server/domain/formQuestion/dto/FormQuestionBaseDto.java
@@ -1,6 +1,6 @@
-package com.kakaotech.team18.backend_server.domain.FormQuestion.dto;
+package com.kakaotech.team18.backend_server.domain.formQuestion.dto;
 
-import com.kakaotech.team18.backend_server.domain.FormQuestion.entity.FieldType;
+import com.kakaotech.team18.backend_server.domain.formQuestion.entity.FieldType;
 import java.util.List;
 
 public interface FormQuestionBaseDto {
