+++ src/main/java/com/kakaotech/team18/backend_server/domain/formQuestion/entity/FieldType.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team18.backend_server.domain.FormQuestion.entity;
+package com.kakaotech.team18.backend_server.domain.formQuestion.entity;
 
 public enum FieldType {
     TEXT,
