+++ src/main/java/com/kakaotech/team18/backend_server/domain/formQuestion/service/FormQuestionService.java
@@ -1,0 +1,9 @@
+package com.kakaotech.team18.backend_server.domain.formQuestion.service;
+
+import com.kakaotech.team18.backend_server.domain.formQuestion.entity.FormQuestion;
+
+import java.util.List;
+
+public interface FormQuestionService {
+    List<FormQuestion> getApplicationFormFieldsById(Long applicationFormId);
+}
